requirements.txt
torch
transformers
Pillow
opencv-python
numpy
accelerate>=0.26.0
pyvips

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
